client/requirements.txt
python_pacman==0.4.1
requests==2.18.4
tqdm==4.19.5
apt==0.7.8
python_dateutil==2.6.1
pacman==0.0.0
